requirements.txt
ragas==0.1.7
wandb==0.16.6
datasets==2.19.1
python-dotenv==1.0.1
langchain==0.1.17
langchain-openai==0.1.6
langchain-community==0.0.37
sentence-transformers==2.7.0
openai==1.26.0